FIxes after code review
