+++ HungarianAlgorithm/Hungarian/IWriteable.cs
@@ -1,0 +1,7 @@
+﻿namespace Hungarian
+{
+    public interface IWriteable
+    {
+        void Write(TextWriter writer);
+    }
+}
