+++ README.md
@@ -11,7 +11,9 @@ Following options are supported:
 - `-n` to specify number of wells
 - `-k` to specify number of houses to which single well can supply water
 
-Order of input arguments does not matter. If either n or k is specified new instance of the problem will be generated and saved to with specified filename. When generating new instance of the problem and either k or n is not specified default value which is equal to 2 will be used. If file with specified name already exists it will be overriden.
+Order of input arguments does not matter. If either n or k is specified new instance of the problem will be generated and saved with the specified filename. When generating new instance of the problem and either k or n is not specified default value which is equal to 2 will be used. If file with specified name already exists it will be overriden.
 
 Full example program call (with generating new instance of problem - 5 wells, 30 houses)
 `./Application.exe -o 2_2.txt -n 5 -k 6`
+
+File with calculated solution is saved in the same folder and the with the same name as input file but with `_output.txt` suffix.
